Fixing variant retrieving when toggle is off
Since now it didn't change if you had your toggle Off or On with false, as both options where returning null when asking for the variant value. It happens that null is not exactly the same as false, this fixes it and now properly returns a false when it's false. This problem also happened when any of your multivariant values was a 0 too